Add new file server/list.go
And move the methods in user.go to list.go.
